requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord==1.0.1
discord.py==1.3.3
flake8==3.8.3
idna==2.9
mccabe==0.6.1
multidict==4.7.5
mypy==0.780
mypy-extensions==0.4.3
Pillow==7.1.2
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.13.0
typed-ast==1.4.1
typing==3.7.4.1
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.4.2
